+++ pp/netlist_from_yaml.py
@@ -44,12 +44,6 @@ components:
             DL: 0
         transformations: mirror_x
 
-connections:
-    - [CP1, E0, arm_bot, W0]
-    - [arm_bot, E0, CP2, E0]
-    - [CP1, E1, arm_top, W0]
-    - [arm_top, E0, CP2, E0]
-
 ports_map:
     W0: [CP1, W0]
     E0: [CP2, W0]
@@ -61,6 +55,12 @@ ports_map:
     E_BOT_1: [arm_bot, E_1]
     E_BOT_2: [arm_bot, E_2]
     E_BOT_3: [arm_bot, E_3]
+
+connections:
+    - [CP1, E0, arm_bot, W0]
+    - [arm_bot, E0, CP2, E0]
+    - [CP1, E1, arm_top, W0]
+    - [arm_top, E0, CP2, E0]
 
 """
 )
@@ -101,12 +101,6 @@ def netlist_from_yaml(file: Union[str, pathlib.Path, IO[Any]]) -> Component:
                     DL: 0
                 transformations: mirror_x
 
-        connections:
-            - [CP1, E0, arm_bot, W0]
-            - [arm_bot, E0, CP2, E0]
-            - [CP1, E1, arm_top, W0]
-            - [arm_top, E0, CP2, E0]
-
         ports_map:
             W0: [CP1, W0]
             E0: [CP2, W0]
@@ -118,6 +112,12 @@ def netlist_from_yaml(file: Union[str, pathlib.Path, IO[Any]]) -> Component:
             E_BOT_1: [arm_bot, E_1]
             E_BOT_2: [arm_bot, E_2]
             E_BOT_3: [arm_bot, E_3]
+
+        connections:
+            - [CP1, E0, arm_bot, W0]
+            - [arm_bot, E0, CP2, E0]
+            - [CP1, E1, arm_top, W0]
+            - [arm_top, E0, CP2, E0]
 
     """
 
@@ -138,7 +138,6 @@ def netlist_from_yaml(file: Union[str, pathlib.Path, IO[Any]]) -> Component:
 
     connections = conf.connections
     ports_map = conf.ports_map
-    ports_map = {"W0": ("CP1", "W0"), "E0": ("CP2", "W0")}
     return netlist_to_component(components, connections, ports_map)